ResumeChatbot/requirements.txt
streamlit
google-generativeai
pdfminer.six
tempfile
